daemon: split out snap configuration support to its own api_*.go files
