+++ daemon/api_test.go
@@ -47,7 +47,6 @@ import (
 	"github.com/snapcore/snapd/overlord/assertstate"
 	"github.com/snapcore/snapd/overlord/assertstate/assertstatetest"
 	"github.com/snapcore/snapd/overlord/auth"
-	"github.com/snapcore/snapd/overlord/configstate/config"
 	"github.com/snapcore/snapd/overlord/healthstate"
 	"github.com/snapcore/snapd/overlord/snapstate"
 	"github.com/snapcore/snapd/overlord/state"
@@ -2614,299 +2613,6 @@ func (s *apiSuite) TestTryChangeConflict(c *check.C) {
 	rsp := trySnap(snapsCmd, req, nil, tryDir, snapstate.Flags{}).(*resp)
 	c.Assert(rsp.Type, check.Equals, ResponseTypeError)
 	c.Check(rsp.Result.(*errorResult).Kind, check.Equals, client.ErrorKindSnapChangeConflict)
-}
-
-func (s *apiSuite) runGetConf(c *check.C, snapName string, keys []string, statusCode int) map[string]interface{} {
-	s.vars = map[string]string{"name": snapName}
-	req, err := http.NewRequest("GET", "/v2/snaps/"+snapName+"/conf?keys="+strings.Join(keys, ","), nil)
-	c.Check(err, check.IsNil)
-	rec := httptest.NewRecorder()
-	snapConfCmd.GET(snapConfCmd, req, nil).ServeHTTP(rec, req)
-	c.Check(rec.Code, check.Equals, statusCode)
-
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Check(err, check.IsNil)
-	return body["result"].(map[string]interface{})
-}
-
-func (s *apiSuite) TestGetConfSingleKey(c *check.C) {
-	d := s.daemon(c)
-
-	// Set a config that we'll get in a moment
-	d.overlord.State().Lock()
-	tr := config.NewTransaction(d.overlord.State())
-	tr.Set("test-snap", "test-key1", "test-value1")
-	tr.Set("test-snap", "test-key2", "test-value2")
-	tr.Commit()
-	d.overlord.State().Unlock()
-
-	result := s.runGetConf(c, "test-snap", []string{"test-key1"}, 200)
-	c.Check(result, check.DeepEquals, map[string]interface{}{"test-key1": "test-value1"})
-
-	result = s.runGetConf(c, "test-snap", []string{"test-key1", "test-key2"}, 200)
-	c.Check(result, check.DeepEquals, map[string]interface{}{"test-key1": "test-value1", "test-key2": "test-value2"})
-}
-
-func (s *apiSuite) TestGetConfCoreSystemAlias(c *check.C) {
-	d := s.daemon(c)
-
-	// Set a config that we'll get in a moment
-	d.overlord.State().Lock()
-	tr := config.NewTransaction(d.overlord.State())
-	tr.Set("core", "test-key1", "test-value1")
-	tr.Commit()
-	d.overlord.State().Unlock()
-
-	result := s.runGetConf(c, "core", []string{"test-key1"}, 200)
-	c.Check(result, check.DeepEquals, map[string]interface{}{"test-key1": "test-value1"})
-
-	result = s.runGetConf(c, "system", []string{"test-key1"}, 200)
-	c.Check(result, check.DeepEquals, map[string]interface{}{"test-key1": "test-value1"})
-}
-
-func (s *apiSuite) TestGetConfMissingKey(c *check.C) {
-	result := s.runGetConf(c, "test-snap", []string{"test-key2"}, 400)
-	c.Check(result, check.DeepEquals, map[string]interface{}{
-		"value": map[string]interface{}{
-			"SnapName": "test-snap",
-			"Key":      "test-key2",
-		},
-		"message": `snap "test-snap" has no "test-key2" configuration option`,
-		"kind":    "option-not-found",
-	})
-}
-
-func (s *apiSuite) TestGetRootDocument(c *check.C) {
-	d := s.daemon(c)
-	d.overlord.State().Lock()
-	tr := config.NewTransaction(d.overlord.State())
-	tr.Set("test-snap", "test-key1", "test-value1")
-	tr.Set("test-snap", "test-key2", "test-value2")
-	tr.Commit()
-	d.overlord.State().Unlock()
-
-	result := s.runGetConf(c, "test-snap", nil, 200)
-	c.Check(result, check.DeepEquals, map[string]interface{}{"test-key1": "test-value1", "test-key2": "test-value2"})
-}
-
-func (s *apiSuite) TestGetConfBadKey(c *check.C) {
-	s.daemon(c)
-	// TODO: this one in particular should really be a 400 also
-	result := s.runGetConf(c, "test-snap", []string{"."}, 500)
-	c.Check(result, check.DeepEquals, map[string]interface{}{"message": `invalid option name: ""`})
-}
-
-func (s *apiSuite) TestSetConf(c *check.C) {
-	d := s.daemon(c)
-	s.MockSnap(c, configYaml)
-
-	// Mock the hook runner
-	hookRunner := testutil.MockCommand(c, "snap", "")
-	defer hookRunner.Restore()
-
-	d.overlord.Loop()
-	defer d.overlord.Stop()
-
-	text, err := json.Marshal(map[string]interface{}{"key": "value"})
-	c.Assert(err, check.IsNil)
-
-	buffer := bytes.NewBuffer(text)
-	req, err := http.NewRequest("PUT", "/v2/snaps/config-snap/conf", buffer)
-	c.Assert(err, check.IsNil)
-
-	s.vars = map[string]string{"name": "config-snap"}
-
-	rec := httptest.NewRecorder()
-	snapConfCmd.PUT(snapConfCmd, req, nil).ServeHTTP(rec, req)
-	c.Check(rec.Code, check.Equals, 202)
-
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Assert(err, check.IsNil)
-	id := body["change"].(string)
-
-	st := d.overlord.State()
-	st.Lock()
-	chg := st.Change(id)
-	st.Unlock()
-	c.Assert(chg, check.NotNil)
-
-	<-chg.Ready()
-
-	st.Lock()
-	err = chg.Err()
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-
-	// Check that the configure hook was run correctly
-	c.Check(hookRunner.Calls(), check.DeepEquals, [][]string{{
-		"snap", "run", "--hook", "configure", "-r", "unset", "config-snap",
-	}})
-}
-
-func (s *apiSuite) TestSetConfCoreSystemAlias(c *check.C) {
-	d := s.daemon(c)
-	s.MockSnap(c, `
-name: core
-version: 1
-`)
-	// Mock the hook runner
-	hookRunner := testutil.MockCommand(c, "snap", "")
-	defer hookRunner.Restore()
-
-	d.overlord.Loop()
-	defer d.overlord.Stop()
-
-	text, err := json.Marshal(map[string]interface{}{"proxy.ftp": "value"})
-	c.Assert(err, check.IsNil)
-
-	buffer := bytes.NewBuffer(text)
-	req, err := http.NewRequest("PUT", "/v2/snaps/system/conf", buffer)
-	c.Assert(err, check.IsNil)
-
-	s.vars = map[string]string{"name": "system"}
-
-	rec := httptest.NewRecorder()
-	snapConfCmd.PUT(snapConfCmd, req, nil).ServeHTTP(rec, req)
-	c.Check(rec.Code, check.Equals, 202)
-
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Assert(err, check.IsNil)
-	id := body["change"].(string)
-
-	st := d.overlord.State()
-	st.Lock()
-	chg := st.Change(id)
-	st.Unlock()
-	c.Assert(chg, check.NotNil)
-
-	<-chg.Ready()
-
-	st.Lock()
-	err = chg.Err()
-	c.Assert(err, check.IsNil)
-
-	tr := config.NewTransaction(st)
-	st.Unlock()
-	c.Assert(err, check.IsNil)
-
-	var value string
-	tr.Get("core", "proxy.ftp", &value)
-	c.Assert(value, check.Equals, "value")
-
-}
-
-func (s *apiSuite) TestSetConfNumber(c *check.C) {
-	d := s.daemon(c)
-	s.MockSnap(c, configYaml)
-
-	// Mock the hook runner
-	hookRunner := testutil.MockCommand(c, "snap", "")
-	defer hookRunner.Restore()
-
-	d.overlord.Loop()
-	defer d.overlord.Stop()
-
-	text, err := json.Marshal(map[string]interface{}{"key": 1234567890})
-	c.Assert(err, check.IsNil)
-
-	buffer := bytes.NewBuffer(text)
-	req, err := http.NewRequest("PUT", "/v2/snaps/config-snap/conf", buffer)
-	c.Assert(err, check.IsNil)
-
-	s.vars = map[string]string{"name": "config-snap"}
-
-	rec := httptest.NewRecorder()
-	snapConfCmd.PUT(snapConfCmd, req, nil).ServeHTTP(rec, req)
-	c.Check(rec.Code, check.Equals, 202)
-
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Assert(err, check.IsNil)
-	id := body["change"].(string)
-
-	st := d.overlord.State()
-	st.Lock()
-	chg := st.Change(id)
-	st.Unlock()
-	c.Assert(chg, check.NotNil)
-
-	<-chg.Ready()
-
-	st.Lock()
-	defer st.Unlock()
-	tr := config.NewTransaction(d.overlord.State())
-	var result interface{}
-	c.Assert(tr.Get("config-snap", "key", &result), check.IsNil)
-	c.Assert(result, check.DeepEquals, json.Number("1234567890"))
-}
-
-func (s *apiSuite) TestSetConfBadSnap(c *check.C) {
-	s.daemonWithOverlordMock(c)
-
-	text, err := json.Marshal(map[string]interface{}{"key": "value"})
-	c.Assert(err, check.IsNil)
-
-	buffer := bytes.NewBuffer(text)
-	req, err := http.NewRequest("PUT", "/v2/snaps/config-snap/conf", buffer)
-	c.Assert(err, check.IsNil)
-
-	s.vars = map[string]string{"name": "config-snap"}
-
-	rec := httptest.NewRecorder()
-	snapConfCmd.PUT(snapConfCmd, req, nil).ServeHTTP(rec, req)
-	c.Check(rec.Code, check.Equals, 404)
-
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Assert(err, check.IsNil)
-	c.Check(body, check.DeepEquals, map[string]interface{}{
-		"status-code": 404.,
-		"status":      "Not Found",
-		"result": map[string]interface{}{
-			"message": `snap "config-snap" is not installed`,
-			"kind":    "snap-not-found",
-			"value":   "config-snap",
-		},
-		"type": "error"})
-}
-
-func (s *apiSuite) TestSetConfChangeConflict(c *check.C) {
-	s.daemon(c)
-	s.MockSnap(c, configYaml)
-
-	s.SimulateConflict("config-snap")
-
-	text, err := json.Marshal(map[string]interface{}{"key": "value"})
-	c.Assert(err, check.IsNil)
-
-	buffer := bytes.NewBuffer(text)
-	req, err := http.NewRequest("PUT", "/v2/snaps/config-snap/conf", buffer)
-	c.Assert(err, check.IsNil)
-
-	s.vars = map[string]string{"name": "config-snap"}
-
-	rec := httptest.NewRecorder()
-	snapConfCmd.PUT(snapConfCmd, req, nil).ServeHTTP(rec, req)
-	c.Check(rec.Code, check.Equals, 409)
-
-	var body map[string]interface{}
-	err = json.Unmarshal(rec.Body.Bytes(), &body)
-	c.Assert(err, check.IsNil)
-	c.Check(body, check.DeepEquals, map[string]interface{}{
-		"status-code": 409.,
-		"status":      "Conflict",
-		"result": map[string]interface{}{
-			"message": `snap "config-snap" has "manip" change in progress`,
-			"kind":    "snap-change-conflict",
-			"value": map[string]interface{}{
-				"change-kind": "manip",
-				"snap-name":   "config-snap",
-			},
-		},
-		"type": "error"})
 }
 
 func (s *apiSuite) TestAppIconGet(c *check.C) {
